@@ -1,14 +1,24 @@
-import {BaseStep} from './Step';
-import {dim} from '../Helper';
-const pj = require('../../package.json');
+import { Answers } from 'inquirer';
+import { BaseStep } from './Step';
+import { dim } from '../Helper';
 
-export default class Initial extends BaseStep {
-  emit() {
-    dim('Starting Sentry setup...');
-    dim(`version: ${pj.version}`);
+let wizardPackage: any = {};
+let projectPackage: any = {};
+
+try {
+  // If we run directly in setup-wizard
+  projectPackage = require('../../package.json');
+} catch {
+  projectPackage = require(`${process.cwd()}/package.json`);
+  wizardPackage = require(`${process.cwd()}/node_modules/@sentry/setup-wizard/package.json`);
+}
+
+export class Initial extends BaseStep {
+  emit(answers: Answers) {
+    dim('Running Sentry Setup Wizard...');
     // TODO: get sentry cli version
     let sentryCliVersion = 'TODO';
-    dim(`sentry-cli version: ${sentryCliVersion}`);
+    dim(`version: ${wizardPackage.version} | sentry-cli version: ${sentryCliVersion}`);
     return Promise.resolve({});
   }
 }
