@@ -1,4 +1,4 @@
-import {setupCordova} from './lib/Setup';
+import { run } from './lib/Setup';
 const argv = require('yargs').boolean('debug').argv;
 
-setupCordova(argv);
+run(argv);
